Move src/include/evict.h src/evict/evict.h
